requirements.txt
fabric3
django
djangorestframework